moves the local storage logic to the BaseUmbracoIndexer
